refactor: 적은 돈으로 1등 당첨시 수익률의 범위가 int를 넘을 수 있어서 long으로 변환

+++ src/main/java/lotto/domain/ResultStatistic.java
@@ -52,12 +52,12 @@ public class ResultStatistic {
 		return this.results;
 	}
 
-	public int calculateRevenueRate(MoneyForLotto moneyForLotto) {
+	public long calculateRevenueRate(MoneyForLotto moneyForLotto) {
 		return (100 * getTotalRevenue() / moneyForLotto.getMoneyForLotto());
 	}
 
-	private int getTotalRevenue() {
-		int totalRevenue = 0;
+	private long getTotalRevenue() {
+		long totalRevenue = 0;
 
 		for (Rank rank : Rank.values()) {
 			totalRevenue += results.get(rank) * rank.getReward();
